requirements.txt
Babel==2.5.1
certifi==2017.11.5
chardet==3.0.4
Django==1.11.7
django-filter==1.1.0
django-geoposition==0.3.0
django-multiselectfield==0.1.8
django-phonenumber-field==1.3.0
djangorestframework==3.7.7
djangorestframework-jwt==1.11.0
idna==2.6
jsonfield==2.0.2
phonenumberslite==8.8.6
Pillow==5.0.0
PyJWT==1.5.3
pytz==2017.3
requests==2.18.4
urllib3==1.22
